added Calculate GPA for each student, also new 2 similar tasks, added today

+++ GPACalculatorAPI/Repositoreis/SubjectRepository.cs
@@ -26,7 +26,7 @@ namespace GPACalculatorAPI.Repositoreis
         {
             var subject = new SubjectEntity();
             subject.Name = request.Name;
-            subject.Credit = request.Credit;
+            subject.Credit = (int)request.Credit;
             await _db.Subjects.AddAsync(subject);
 
             return subject;
